Versioning: Update version tag from GIT repo
This commit will update the version field of the generated deb package from the git repo version tag.
the tag needs to be in the format "vyos/<version>" eg. "vyos/1.2.5" 1.2.5 is then used as the version
The version field will be one of the following syntaxes:
On the tagged commit:
    - <version>
    eg. 1.2.5

not on a tagged commit:
    - <version>-<commits from tag>-g<commit id>
    eg. 1.2.5-4-g23232343

on a commit with unstaged changes:
    - <verson>-dirty
    eg. 1.2.5-dirty
    - <version>-<commits from tag>-g<commit id>-dirty
    eg. 1.2.5-4-g23232343-dirty

This will make it clear what state the repo was in when the package was generated.

if no valid tag is recieved from `git describe`, the script uses the version noted in the changelog
file.

This script also only updates the package version tag, the changelog file and .changes files are not
altered using this process